pyproject.toml
[tool.poetry]
name = "consistent_depth"
version = "0.1.0"
description = "Windows adaptation for consistence depth from Facebook"
authors = ["Didier Bröska <didier.beheydt@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = [
    "AI",
    "Depth map",
    "video"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development",
]

[tool.poetry.dependencies]
python = "^3.8"
ffmpeg-python = "^0.2.0"
Pillow = "^8.0.1"
numpy = "^1.19.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
isort = "^5.6.4"
mypy = "^0.790"
pytest = "^6.1.2"
black = "^20.8b1"
safety = "^1.9.0"
gdown = "^3.12.2"
pytest-cov = "^2.10.1"

[tool.pytest.ini_options]
addopts = "-ra --capture=tee-sys"
testpaths = "tests/"

[tool.isort]
profile = "black"
include_trailing_comma="True"
force_grid_wrap="0"
use_parentheses="True"
line_length="79"

[tool.black]
line-length="79"
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38, py39
skipsdist = true
isolated_build = true
[tox:.package]
basepython = python3
[testenv]
whitelist_externals = poetry
commands =
    poetry install -v
    poetry run python --version
    poetry run safety check
    poetry run pytest tests/
"""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
